Trying to make a scheduler

+++ src/klibc/math.c
@@ -1,7 +1,7 @@
 #include "math.h"
 
-uint64_t abs(int64_t in) {
+int64_t abs(int64_t in) {
     /* If input is less than 0, return the opposite of the input, otherwise return input */
-    if (in < 0) return (uint64_t) -in;
-    return (uint64_t) in;
+    if (in < 0) return -in;
+    return in;
 }
